Rearrange preprocessor routines to handle implicit macro concatenation
If a macro usage comes right after a normal token (with no whitespace in-between), the first expanded token might need to be concatenated onto the normal token. This gets handled basically for free with a traditional preprocessor, but is a bit more trouble for us because we use a token-based preprocessor to avoid re-lexing every file twice.