Refactoring of the symbol hierarchy
This reorganizes scope symbols so that they always create all of their members in one go and then don't mutate afterwards. This was always the original plan but it got pretty complicated along the way.

A nice side effect of this change is that all symbols hold their member maps by non-owning pointer, so we don't have to worry about their lifetimes anymore (can just throw them in the BumpAllocator and be done with it.)