Clean up and consolidate type - symbol ownership
This gets rid of the Lazy types and consolidates them into a new type called DeclaredType. All ValueSymbols now have a declared type, lifting a lot of disparate functionality out of the member symbols. This makes everything easier to follow and understand, fixes a few bugs, and makes it possible to always know the declared syntax for a type.