Refactor diagnostic reporting system
Split DiagnosticWriter into two separate components, one called DiagnosticEngine that controls whether and how diagnostics get reported, and the other called DiagnosticClient which is an interface for swapping out how a particular diagnostic actually gets rendered.

The engine has new options added (not yet hooked up) that control things like which warnings should be ignored, which should be treated as errors, the maximum number of errors to report, etc.

There's only one client implemented for now, TextDiagnosticClient, which does what the old system used to do and renders the diagnostic as a text string fit for printing to the console.