Clean up diagnostic filtering
Previously the compilation tracked a list of all diagnostics and at the end would do a bunch of work to collapse them down into a sane set to display to the user. The problem was that filtering only got applied at the end, but the errorLimit feature would kick in before filtering, meaning you could, in extreme cases, end up filtering out all of the errors you had discovered and have nothing to show.

Now we do filtering and coalescing on the fly, which makes the error limiter more accurate. It should hopefully be a bit more efficient as well. This also let me get rid of the error counter tracking in the Diagnostics objects, which is a nice win.