Time.cpp: include <cmath>
Time.cpp uses `std::round` from <cmath>. When SLANG_USE_BOOST is on,
cmath is pulled in somehow from a transitive include. When
SLANG_USE_BOOST is off, at least on some systems, this include
is not pulled in.

Include <cmath> explicitly in Time.cpp to ensure `std::round` is
available, including for certain non-boost-using builds.