New slang-tidy checks (#816)
* New slang-tidy check: XilinxDoNotCareValues

This new checks reports usage of do-not-care values with the syntax
x'd?. Which Xilinx tools do not treat it as x'b? and it could be a
source of bugs.

* New slang-tidy check: CastSignedIndex

This check reports the use of casting signed indexes that are use as
selectors of an array. This cast can produce negative indexes, making
some positions of the array inaccessible