slang-reflect fixes on data assignment and namespaces (#844)
* [slang-reflect] Correctly assign members greater than 64 bits

We were not checking if a struct member was greater than 64 bits
when giving it a value in the struct constructor, and always using the
to_uint64() function loosing data in the conversion.

* [slang-reflect] Provide type namespace if necessary in get_* functions

When returning a non-primitive type from a get_* function, we were
assuming that the type was part of the current namespace, which is not
always true. This commit fixes this issue by adding the namespace
when necessary.