Fix slang-tidy synthesis checks
* slang-tidy fix OnlyAssignedOnReset and RegisterHasNoReset:
This commit will ensure that both checks will no longer be ignoring the
configuration setting 'resetIsActiveHigh' and the possible negation of
the reset signal. Two test cases have also been added to consider this
bug.

* The error location of the warnings generated by the
OnlyAssignedOnReset and RegisterHasNoReset are now more precise and
target the variable instance where the error is found instead of the
variable declaration.