Add multiple slang-tidy checks and minor fixes in existing ones (#1040)
* New slang-tidy check: AlwaysCombNamed

This new check enforces that each always_comb block is named

* New slang-tidy check: GenerateNamed

This new check enforces that each generate block is named.
This also is applied to the for block

* New slang-tidy check: NoDotVarInPortConnection

This check ensures that no module instantiation uses the sintax '.var' in
the port connection, and ensures that '.var(var)' is used instead

* Fix in the warning location of a check

Now, when '.*' is used with the NoDotStarInPortConnection check enabled,
the warning will be located where the '.*' character is, instead of the
location of the module that was being called.

* New slang-tidy check: NoLegacyGenerate

This new check enforces that no generate block is used in the code,
since it is an old system verilog syntax

* New check: AlwaysFFAssignmentOusideConditional

This new synthesis check ensures that if there is a conditional
block with reset, all the assignments must be either inside the
if or else statements.

* New slang-tidy check: UnusedSensitiveSignal

This new check verifies that every signal inside the sensitivity list
of an 'always' or 'always_ff' block is used inside the block
statement. The only exception is the clock signal

* Fix slang-tidy synthesis checks

* slang-tidy fix OnlyAssignedOnReset and RegisterHasNoReset:
This commit will ensure that both checks will no longer be ignoring the
configuration setting 'resetIsActiveHigh' and the possible negation of
the reset signal. Two test cases have also been added to consider this
bug.

* The error location of the warnings generated by the
OnlyAssignedOnReset and RegisterHasNoReset are now more precise and
target the variable instance where the error is found instead of the
variable declaration.

* style: pre-commit fixes

* Remove redundant condition in slang-tidy