disable MSAN instrumentation for SmallVector destructor
Because of the destruction order when handling small vectors, MSAN is
throwing a `use-of-uninitialized-value` error. The best explantation
we came up with is that:
- first `SmallVector` releases the stackBase memory
- MSAN poisons the freed memory
- the `SmallVectorBase` destructor calls cleanup, which does a check
  for `isSmall()` which accesses the poisoned memory locations.

This "fix" essentially creates an explicit destructor for SmallVector
which is annotated to disable MSAN checks because we are quite
confident that the underlying memory released is still available for
the cleanup check.